Refactoring modèles graphique
