+++ Sudoku.GraphicalModelSolver/GraphicalModelIterativeSolver.cs
@@ -1,0 +1,207 @@
+﻿using System;
+using System.Linq;
+using Microsoft.ML.Probabilistic.Distributions;
+using Microsoft.ML.Probabilistic.Math;
+using Microsoft.ML.Probabilistic.Models;
+using Sudoku.Shared;
+
+namespace GraphicModelSolver
+{
+
+    public class GraphicalModelIterativeSolver : GraphicalModelRobustSolver
+    {
+
+        public int NbIterationCells { get; set; } = 2;
+
+
+
+        protected override void DoInference(Dirichlet[] dirArray, int[] sudokuCells)
+        {
+
+            int cellDiscovered = sudokuCells.Count(c =>c > 0 );
+
+            // Iteration tant que l'on a pas découvert toutes les cases
+            while (cellDiscovered < CellIndices.Count - 1)
+            {
+
+                Dirichlet[] cellsProbsPosterior = InferenceEngine.Infer<Dirichlet[]>(ProbCells);
+
+                Dirichlet[] bestCellsProbsPosterior = getBestDirichletSubArray(cellsProbsPosterior, NbIterationCells, sudokuCells);
+
+                double minDiricheltValue = 1;
+                // Récupère le plus petit mode.Max() des meilleures bestCellsProbsPosterior 
+                foreach (var dirichlet in bestCellsProbsPosterior)
+                {
+                    var mode = dirichlet.GetMode();
+
+
+                    if (mode.Max() <= minDiricheltValue)
+                    {
+                        minDiricheltValue = mode.Max();
+                    }
+                }
+
+                //todo: pourquoi reparcourir les cellules alors que nous avons déjà la liste de celles qui nous intéressent?
+                foreach (var cellIndex in CellIndices)
+                {
+                    if (sudokuCells[cellIndex] == 0)
+                    {
+                        var mode = cellsProbsPosterior[cellIndex].GetMode();
+
+                        if (mode.Max() >= minDiricheltValue)
+                        {
+                            var value = mode.IndexOf(mode.Max()) + 1;
+                            Vector v = Vector.Constant(CellDomain.Count, EpsilonProba);
+                            v[value - 1] = FixedValueProba;
+
+                            dirArray[cellIndex] = Dirichlet.PointMass(v);
+
+                            sudokuCells[cellIndex] = value;
+                            cellDiscovered++;
+                        }
+                    }
+                }
+
+                CellsPrior.ObservedValue = dirArray;
+            }
+        }
+
+        private Dirichlet[] getBestDirichletSubArray(Dirichlet[] dirichletArray, int N, int[] sudokuCells)
+        {
+            // Initialise la liste des N meilleurs Dirichlet avec les N premiers Dirichlet de dirichletArray
+            Dirichlet[] bestDir = new Dirichlet[N];
+
+
+            //Todo: cette initialisation ne revient-elle pas à chercher la première cellule vide N fois?
+            for (int i = 0; i < N; i++)
+            {
+                foreach (var cellIndex in CellIndices)
+                {
+                    if (sudokuCells[cellIndex] == 0)
+                    {
+                        bestDir[i] = dirichletArray[cellIndex];
+                        break;
+                    }
+                }
+            }
+
+            // Pour chaque cellule == 0 du sudoku
+            foreach (var cellIndex in CellIndices)
+            {
+                if (sudokuCells[cellIndex] == 0)
+                {
+                    var mode = dirichletArray[cellIndex].GetMode();
+                    Dirichlet minDir = bestDir[0];
+
+                    // Récupère le Dirichlet le plus petit de la liste des meilleurs Dirichlet
+                    foreach (var dir in bestDir)
+                    {
+                        var dirMode = dir.GetMode();
+                        var minDirMode = minDir.GetMode();
+
+                        if (dirMode.Max() < minDirMode.Max())
+                        {
+                            minDir = dir;
+                        }
+                    }
+                    // Remplace ce Dirichlet si la valeurs max du Dirichlet de la cellule actuelle est supèrieur
+                    if (minDir.GetMode().Max() < mode.Max())
+                    {
+                        bestDir[Array.IndexOf(bestDir, minDir)] = dirichletArray[cellIndex];
+                    }
+                }
+            }
+            return bestDir;
+        }
+
+    }
+
+
+
+
+
+    /// <summary>
+    /// Tentative de redéfinition de la méthode itérative combinant des simplifications diverses et une nouvelle définition de la confiance.
+    /// </summary>
+    public class GraphicalModelAlernateIterativeSolver : GraphicalModelIterativeSolver
+    {
+        protected override void DoInference(Dirichlet[] dirArray, int[] sudokuCells)
+        {
+            int cellDiscovered = sudokuCells.Count(c => c > 0);
+
+            // Iteration tant que l'on a pas découvert toutes les cases
+            while (cellDiscovered < CellIndices.Count - 1)
+            {
+
+                Dirichlet[] cellsProbsPosterior = InferenceEngine.Infer<Dirichlet[]>(ProbCells);
+
+                var bestCells = GetBestCondidenceSubArray(cellsProbsPosterior, NbIterationCells, sudokuCells);
+
+
+                foreach (var cellPair in bestCells)
+                {
+
+                    var mode = cellsProbsPosterior[cellPair.cellIndex].GetMode();
+
+
+                    var value = mode.IndexOf(mode.Max()) + 1;
+                    Vector v = Vector.Constant(CellDomain.Count, EpsilonProba);
+                    v[value - 1] = FixedValueProba;
+
+                    dirArray[cellPair.cellIndex] = Dirichlet.PointMass(v);
+
+                    sudokuCells[cellPair.cellIndex] = value;
+                    cellDiscovered++;
+                }
+
+                CellsPrior.ObservedValue = dirArray;
+            }
+        }
+
+
+        private (int cellIndex, double confidence)[] GetBestCondidenceSubArray(Dirichlet[] dirichletArray, int nbBests, int[] sudokuCells)
+        {
+            var toReturn = new (int cellIndex, double confidence)[nbBests];
+            var minIndex = nbBests - 1;
+
+            foreach (var cellIndex in CellIndices)
+            {
+                if (sudokuCells[cellIndex] == 0)
+                {
+                    var confidence = GetDirichletConfidence(dirichletArray[cellIndex]);
+                    if (confidence > toReturn[minIndex].confidence)
+                    {
+                        toReturn[minIndex] = (cellIndex, confidence);
+                        // recompute minIndex
+                        minIndex = 0;
+                        for (int i = 0; i < toReturn.Length; i++)
+                        {
+                            if (toReturn[i].confidence < toReturn[minIndex].confidence) { minIndex = i; }
+                        }
+                    }
+                }
+            }
+            return toReturn;
+        }
+
+
+
+        /// <summary>
+        /// Provides a confidence index by the difference between best and second best probability 
+        /// </summary>
+        private double GetDirichletConfidence(Dirichlet distribution)
+        {
+            var dMode = distribution.GetMode();
+            // Récupération des deux valeurs max
+            var maxValues = dMode.Aggregate<double, (double max1, double max2)>((dMode[0], dMode[0]), (currentMaxes, val) =>
+                val > currentMaxes.max1 ? (val, currentMaxes.max2) : val > currentMaxes.max2 ? (currentMaxes.max1, val) : currentMaxes);
+            return maxValues.max1 - maxValues.max2;
+        }
+    }
+
+
+
+
+
+
+}
